Updated Instructions
Create a web application that displays a listing of products with a category select
dropdown. By default, the first category should be selected and the
products of that category should be shown. Users can change the category of the product which should fetch the list of products from the
backend service based on the selected category. Upon clicking on any of the products, the app should navigate to the details page of the product.

    npm install

In the project directory, you can run:

Runs the app in the development mode.\
Open [http://localhost:3000](http://localhost:3000) to view it in the browser.